format gravity fsm
